# backend/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "compiler.Result": {
                "properties": {
                    "log": {
                        "type": "string"
                    },
                    "pdfAvailable": {
                        "type": "boolean"
                    },
                    "success": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "project.CreateProjectRequest": {
                "properties": {
                    "bibliography": {
                        "type": "string"
                    },
                    "engine": {
                        "type": "string"
                    },
                    "mainFile": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "project.Project": {
                "properties": {
                    "bibliography": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "engine": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "mainFile": {
                        "type": "string"
                    },
                    "memberships": {
                        "items": {
                            "$ref": "#/components/schemas/project.ProjectMembership"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "name": {
                        "type": "string"
                    },
                    "ownerId": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "SessionCookie": {
                "description": "HTTP-only session cookie used to authenticate requests.",
                "in": "cookie",
                "name": "paper_server_session",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Returns the current API server health status.",
        "title": "EnderTex API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/projects": {
            "get": {
                "description": "Returns projects available to the authenticated user.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/project.Project"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "summary": "List projects",
                "tags": [
                    "Projects"
                ]
            },
            "post": {
                "description": "Creates a new LaTeX project and assigns the authenticated user as its owner.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/project.CreateProjectRequest",
                                "summary": "request",
                                "description": "Project configuration"
                            }
                        }
                    },
                    "description": "Project configuration",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/project.Project"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "summary": "Create project",
                "tags": [
                    "Projects"
                ]
            }
        },
        "/projects/{projectId}": {
            "get": {
                "description": "Returns project metadata and membership information for a project.",
                "parameters": [
                    {
                        "description": "Project ID",
                        "in": "path",
                        "name": "projectId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/project.Project"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Get project",
                "tags": [
                    "Projects"
                ]
            }
        },
        "/projects/{projectId}/compile": {
            "post": {
                "description": "Compiles the project's main LaTeX document using the configured LaTeX engine.",
                "parameters": [
                    {
                        "description": "Project ID",
                        "in": "path",
                        "name": "projectId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/compiler.Result"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Invalid project path"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Compilation or server error"
                    }
                },
                "summary": "Compile project",
                "tags": [
                    "Compilation"
                ]
            }
        },
...
    },
    "openapi": "3.0.0",
    "servers": [
        {
            "url": "/api"
        }
    ]
}